pyproject.toml
[project]
name = "hud-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.3.0",
    "mcp[cli]>=1.10.1",
    "pillow",
    "pydantic>=2.11.4",
    "numpy>=1.26.0",
    "pyyaml>=6.0.2",
    "pymongo",
    "anthropic>=0.49.0",
    "jsonschema>=4.21.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.28.0",
    "imageio>=2.31.0",
    "pytest>=8.4.0",
    "packaging>=21.0",
    "defusedxml>=0.7.1",
    "hud-python[agents]>=0.4.74",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.11.2", "pre-commit>=3.5.0"]

[project.scripts]
hud_eval = "hud_controller.app:main"
# Same as `hud` CLI but patches grouped eval so "Mean pass%" = mean cocotb pass fraction (not reward>0 count).
hud-patched = "hud_controller.hud_cli_entry:main"
setup_problem = "hud_controller.app:setup_problem_script"
grade_problem = "hud_controller.app:grade_problem_script"
validate_problem = "hud_controller.app:validate_problem_script"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "I",  # isort
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.hatch.metadata]
allow-direct-references = true
